refactor(projects): organize appearance code
