+++ packages/app/src/styles/theme.ts
@@ -638,3 +638,16 @@ export const THEME_SWATCHES: Record<ThemeName, string> = {
   claude: "#D97757",
   ghostty: "#8caaee",
 };
+
+export const PROJECT_ICON_COLORS = [
+  "#8b5cf6",
+  "#0ea5e9",
+  "#10b981",
+  "#f97316",
+  "#ec4899",
+  "#6366f1",
+  "#14b8a6",
+  "#ef4444",
+  "#eab308",
+  "#3b82f6",
+] as const;
