+++ packages/protocol/src/project-appearance.ts
@@ -1,0 +1,1 @@
+export const MAX_PROJECT_ICON_TEXT_CHARACTERS = 8;
